chatbot-function/requirements.txt
langgraph==1.0.1
functions-framework==3.10.0
langchain-google-vertexai==2.1.2
langgraph-checkpoint-firestore==0.1.7
langchain-google-firestore==0.5.0
langchain-community==0.3.31
langsmith==0.4.43
langchain-core==0.3.79
pydantic==2.12.4
pydantic-settings==2.12.0
python-dotenv==1.2.1
tenacity==9.1.2
google-cloud-firestore==2.21.0
google-cloud-logging==3.12.1
google-auth==2.43.0
firebase-admin==6.9.0
python-json-logger==4.0.0
tavily==1.1.0